User create (#4)
* User Create screen

-generates user ID
-Hashes password
-Writes user to JSON file

* Proper OOP, split off hash and JSON parse/write.

1. Split off the password hasher and JSON parse/write into their own
files

2. made the user creator actually output well-formed JSON, with a
decently usable structure (I hope)

3. Switched from org.json.simple to GSON

NOTE: The read/parse functions need to have something in the file to
read from. I use an empty json object, and unfortuately at this point
that empty object stays with the file. I'm working on fixing that though

* Cleaned up

Cleaned up a few things from the last commit. No actual code has changed

* Made the json file with accounts searchable by username

* searchable arrays

Gives our json arrays names, so that when we read them again, we can search for things by their names

* Fixed Array

Fixed the bug in JSONParseWrite so that arrays are actually searchable. This closes my previous issue from today

* Put this in the wrong spot

* Fixed Arrays

Fixed the arrays so that they are no longer the weird nested abomination that they were before, RE: #3 . This also means the arrays are now searchable by username, which should make it possible to actually log into the system eventually
This closes that issue and makes this branch ready to be pulled in to master.